pyproject.toml
[project]
name = "data-analytics-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
